can now post messages as kevlar bot instead of usr
Separate User and Bot token scopes needed to be added for different
tasks. The user token allows a privileged user the ability to delete
messages, however messages posted will look like the bot. To circumvent
not being able to do this with only a bot token, the bot token is used with
custom message parameters each time the bot needs to post messages. The
bot token overrides the user token when specified.